Setup for React-Redux
