+++ public/vanillajs-index.html
@@ -11,9 +11,18 @@
     />
     <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-    <title>React Redux</title>
+    <title>Vanilla Redux</title>
   </head>
   <body>
-    <div id="root"></div>
+    <button id="add">Add</button>
+    <span>0</span>
+    <button id="minus">Minus</button>
+    <hr />
+    <h1>To Dos</h1>
+    <form>
+      <input type="text" placeholder="Write to do" required />
+      <button>Add</button>
+    </form>
+    <ul></ul>
   </body>
 </html>
