+++ src/components/App.js
@@ -1,0 +1,14 @@
+import { HashRouter as Router, Route, Routes } from "react-router-dom";
+import Detail from "../routes/Detail";
+import Home from "../routes/Home";
+
+export default function App() {
+  return (
+    <Router>
+      <Routes>
+        <Route path="/" exact element={<Home />}></Route>
+        <Route path="/:id" element={<Detail />}></Route>
+      </Routes>
+    </Router>
+  );
+}
