+++ src/routes/Home.js
@@ -1,0 +1,26 @@
+import { useState } from "react";
+
+export default function Home() {
+  const [text, setText] = useState("");
+
+  const onChange = (e) => {
+    setText(e.target.value);
+  };
+
+  const onSubmit = (e) => {
+    e.preventDefault();
+    console.log(text);
+    setText("");
+  };
+
+  return (
+    <>
+      <h1>To Do</h1>
+      <form onSubmit={onSubmit}>
+        <input type="text" value={text} onChange={onChange} required />
+        <button>Add</button>
+      </form>
+      <ul></ul>
+    </>
+  );
+}
